Board tiles are now a separate class rather than a structure.
