+++ Assets/Scripts/Battle/BoardTile.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5cf3e3d27dca449ca89856cb243ec27e
+timeCreated: 1492695212
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
