+++ Assets/Scripts/ShipPlacer.cs
@@ -180,8 +180,8 @@ public class ShipPlacer : MonoBehaviour
 
         //processedShip.gameObject.SetActive(showShip); //Shows the ship
         //processedShip.transform.position = player.board.tiles[(int)selectedPositions[Mathf.CeilToInt((float)processedShip.length / 2f) - 1].x, (int)selectedPositions[Mathf.CeilToInt((float)processedShip.length / 2f) - 1].y].worldPosition - Vector3.up * (GameController.playerBoardElevation - 0.4f); //Sets up the position of the ship
-        Vector3 position1 = player.board.tiles[(int)selectedPositions[0].x, (int)selectedPositions[0].y].worldPosition;
-        Vector3 position2 = player.board.tiles[(int)selectedPositions[selectedPositions.Count - 1].x, (int)selectedPositions[selectedPositions.Count - 1].y].worldPosition;
+        Vector3 position1 = player.board.tiles[(int)selectedPositions[0].x, (int)selectedPositions[0].y].transform.position;
+        Vector3 position2 = player.board.tiles[(int)selectedPositions[selectedPositions.Count - 1].x, (int)selectedPositions[selectedPositions.Count - 1].y].transform.position;
 
         processedShip.boardPosition = position1 + (position2 - position1) / 2f - Vector3.up * (GameController.playerBoardElevation - GameController.seaLevel);
         if ((selectedPositions[0] - selectedPositions[1]).x != 0) //Rotates the ship correctly
